Dup the job spec when enqueueing a job
Sidekiq::Client adds the JID it generates to the spec, since the spec is
kept in memory by the scheduler, the next time the job is enqueued the
spec includes this JID and Sidekiq will use that instead of generating a
new one.

This can result in multiple jobs with the same id either enqueued or
active. This breaks the locking mechanism of sidekiq-unique-jobs and
might have some other unwanted effects.